Cleanup random selection function
This still uses the previous random number selection method, it's just been cleaned up to make it more readable and to avoid duplication.

As far as I could tell, there was no reason for the `C101_KinbakuClub_Slaves_Random#` variables to be global. So that's been incorperated into the function to keep things cleaner.