Changes to visiting KC (#60)
* More SlaveTwin

* To cut or not to cut

* SlaveTwin Beta Ready

The SlaveTwin section should be ready for some testing, but the accompanying RopeGroup section still needs a lot of work doing to it.

* Cleanup random selection function

This still uses the previous random number selection method, it's just been cleaned up to make it more readable and to avoid duplication.

As far as I could tell, there was no reason for the `C101_KinbakuClub_Slaves_Random#` variables to be global. So that's been incorperated into the function to keep things cleaner.

* RopeGroup tidied for Alhpa

* Change going to K Club.

* More KC fitting in

* Can leave KC

Changed the door in the club, player can now leave if not in bondage.